provider/kubernetes Remove cassandra as a dependency

+++ experimental/kubernetes/scripts/cleanup-config.sh
@@ -1,8 +1,10 @@
-kubectl delete secret kube-config --namespace=spinnaker
 kubectl delete secret spinnaker-config --namespace=spinnaker
+kubectl delete secret creds-config --namespace=spinnaker
 
 CONF_DIR=../../config/
 
 for FILENAME in $CONF_DIR*.yml; do
-    rm config/${FILENAME:${#CONF_DIR}}
+    if [ ${FILENAME:${#CONF_DIR}} != "default-spinnaker-local.yml" ]; then
+        rm config/${FILENAME:${#CONF_DIR}}
+    fi
 done
